refresh src/

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -1,5 +1,7 @@
 package ru.netology.stats;
 
+//    1
+
 public class StatsService {
     public long sum(long[] purchases) {
         long sum = 0;
@@ -11,16 +13,17 @@ public class StatsService {
         return sum;
     }
 
-
+//    2
 
     public long average(long[] purchases) {
         long sum = sum(purchases);
         return sum / purchases.length;
     }
 
+//    3
 
-    public int maxSales(long[] purchases) {
-        int maxMonthSales = (int) purchases[0];
+    public long maxSales(long[] purchases) {
+        long maxMonthSales = purchases[0];
         int monthMaxSales = 0;
         int count = 0;
         for (long purchase : purchases) {
@@ -33,8 +36,10 @@ public class StatsService {
         return monthMaxSales;
 }
 
-    public int minSales(long[] purchases) {
-        int minMonthSales = (int) purchases[0];
+//    4
+
+    public long minSales(long[] purchases) {
+        long minMonthSales = purchases[0];
         int monthMinSales = 0;
         int count = 0;
         for (long purchase : purchases) {
@@ -47,8 +52,10 @@ public class StatsService {
         return monthMinSales;
     }
 
-    public int monthsSalesBelowAverage(long[] purchases) {
-        int average = (int) average(purchases);
+//    5
+
+    public long monthsSalesBelowAverage(long[] purchases) {
+        long average = average(purchases);
         int monthsSalesBelowAverage = 0;
         int count = 0;
         for (long purchase : purchases) {
@@ -60,8 +67,10 @@ public class StatsService {
         return monthsSalesBelowAverage;
     }
 
-    public int monthsSalesOverAverage(long[] purchases) {
-        int average = (int) average(purchases);
+//    6
+
+    public long monthsSalesOverAverage(long[] purchases) {
+        long average = average(purchases);
         int monthSalesOverAverage = 0;
         int count = 0;
         for (long purchase : purchases) {
